--- app/src/main/java/cmpt276/project/model/Mode.java
@@ -1,9 +1,0 @@
-package cmpt276.project.model;
-
-/**
- * MODE
- * contains game modes
- */
-public enum Mode {
-    IMAGES, WORD_IMAGES, EASY, NORMAL, HARD
-}
