--- app/src/main/java/cmpt276/project/model/flickr/GalleryItem.java
@@ -1,43 +1,0 @@
-package cmpt276.project.model.flickr;
-
-import androidx.annotation.NonNull;
-
-/**
- * GALLERY ITEM CLASS
- * Contains functions related to information about each image
- */
-public class GalleryItem {
-    private String mCaption;
-    private String mId;
-    private String mUrl;
-
-    public String getCaption() {
-        return mCaption;
-    }
-
-    public void setCaption(String caption) {
-        mCaption = caption;
-    }
-
-    public String getId() {
-        return mId;
-    }
-
-    public void setId(String id) {
-        mId = id;
-    }
-
-    public String getUrl() {
-        return mUrl;
-    }
-
-    public void setUrl(String url) {
-        mUrl = url;
-    }
-
-    @NonNull
-    @Override
-    public String toString() {
-        return mCaption;
-    }
-}
