--- app/src/main/java/cmpt276/project/model/flickr/QueryPreferences.java
@@ -1,25 +1,0 @@
-package cmpt276.project.model.flickr;
-
-import android.content.Context;
-import android.preference.PreferenceManager;
-
-/**
- * QUERY PREFERENCES
- * Search queries in flickr
- */
-public class QueryPreferences {
-
-    private static final String PREF_SEARCH_QUERY = "searchQuery";
-
-    public static String getStoredQuery(Context context) {
-        return PreferenceManager.getDefaultSharedPreferences(context)
-                .getString(PREF_SEARCH_QUERY, null);
-    }
-
-    public static void setStoredQuery(Context context, String query) {
-        PreferenceManager.getDefaultSharedPreferences(context)
-                .edit()
-                .putString(PREF_SEARCH_QUERY, query)
-                .apply();
-    }
-}
