--- app/src/main/java/cmpt276/project/ui/flickr/SingleFragmentActivity.java
@@ -1,39 +1,0 @@
-package cmpt276.project.ui.flickr;
-
-import android.os.Bundle;
-import androidx.annotation.LayoutRes;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.appcompat.app.AppCompatActivity;
-
-import cmpt276.project.R;
-
-/**
- * SINGLE FRAGMENT CLASS
- * Creates the basic structure of the flickr fragment
- */
-public abstract class SingleFragmentActivity extends AppCompatActivity {
-
-    protected abstract Fragment createFragment();
-
-    @LayoutRes
-    protected int getLayoutResId() {
-        return R.layout.activity_fragment;
-    }
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(getLayoutResId());
-
-        FragmentManager fm = getSupportFragmentManager();
-        Fragment fragment = fm.findFragmentById(R.id.frame_fragment_container);
-
-        if (fragment == null) {
-            fragment = createFragment();
-            fm.beginTransaction()
-                    .add(R.id.frame_fragment_container, fragment)
-                    .commit();
-        }
-    }
-}
