+++ app/src/main/java/project/findit/model/Mode.java
@@ -1,0 +1,9 @@
+package project.findit.model;
+
+/**
+ * MODE
+ * contains game modes
+ */
+public enum Mode {
+    IMAGES, WORD_IMAGES, EASY, NORMAL, HARD
+}
