+++ app/src/main/java/project/findit/model/flickr/QueryPreferences.java
@@ -1,0 +1,25 @@
+package project.findit.model.flickr;
+
+import android.content.Context;
+import android.preference.PreferenceManager;
+
+/**
+ * QUERY PREFERENCES
+ * Search queries in flickr
+ */
+public class QueryPreferences {
+
+    private static final String PREF_SEARCH_QUERY = "searchQuery";
+
+    public static String getStoredQuery(Context context) {
+        return PreferenceManager.getDefaultSharedPreferences(context)
+                .getString(PREF_SEARCH_QUERY, null);
+    }
+
+    public static void setStoredQuery(Context context, String query) {
+        PreferenceManager.getDefaultSharedPreferences(context)
+                .edit()
+                .putString(PREF_SEARCH_QUERY, query)
+                .apply();
+    }
+}
