+++ app/src/main/java/project/findit/ui/flickr/PhotoGalleryActivity.java
@@ -1,0 +1,32 @@
+package project.findit.ui.flickr;
+
+import android.content.Context;
+import android.content.Intent;
+import android.view.MenuItem;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+
+/**
+ * PHOTO GALLERY
+ * This activity calls the Photogallery fragment
+ */
+public class PhotoGalleryActivity extends SingleFragmentActivity {
+
+    public static Intent makeLaunchIntent(Context context){
+        return new Intent(context, PhotoGalleryActivity.class);
+    }
+
+    @Override
+    protected Fragment createFragment() {
+        return PhotoGalleryFragment.newInstance();
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+        if(item.getItemId() == android.R.id.home){
+            onBackPressed();
+        }
+        return super.onOptionsItemSelected(item);
+    }
+}
