+++ app/src/main/res/layout/fragment_photo_gallery.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="cmpt276.project.ui.flickr.PhotoGalleryActivity">
+    tools:context=".ui.flickr.PhotoGalleryActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_photo"
